refactor: replace DefaultBookingFields with RequiredBookingFields and move isInt utility function
